fix: remove broad permissions:read-all, extract duplicated test helper
