+++ .github/workflows/scorecard.yml
@@ -6,14 +6,13 @@ on:
   push:
     branches: [main]
 
-permissions: read-all
-
 jobs:
   scorecard:
     runs-on: ubuntu-latest
     permissions:
-      security-events: write
-      id-token: write
+      security-events: write  # upload SARIF to GitHub Security tab
+      id-token: write         # NOSONAR — required by ossf/scorecard-action for OIDC provenance signing
+      contents: read
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
         with:
